+++ bundles/org.openhab.binding.energenie/src/main/java/org/openhab/binding/energenie/internal/handler/EnergeniePWMHandler.java
@@ -19,12 +19,8 @@ import java.nio.charset.StandardCharsets;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
 
-import javax.measure.Unit;
-
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.eclipse.smarthome.core.library.types.QuantityType;
-import org.eclipse.smarthome.core.library.unit.SmartHomeUnits;
 import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingStatus;
@@ -32,9 +28,8 @@ import org.eclipse.smarthome.core.thing.ThingStatusDetail;
 import org.eclipse.smarthome.core.thing.binding.BaseThingHandler;
 import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.RefreshType;
-import org.eclipse.smarthome.core.types.State;
-import org.eclipse.smarthome.core.types.UnDefType;
 import org.eclipse.smarthome.io.net.http.HttpUtil;
+import org.openhab.binding.energenie.internal.EnergeniePWMStateEnum;
 import org.openhab.binding.energenie.internal.config.EnergenieConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -47,75 +42,6 @@ import org.slf4j.LoggerFactory;
 
 @NonNullByDefault
 public class EnergeniePWMHandler extends BaseThingHandler {
-
-    public enum PWMState {
-        VOLTAGE("var V  = ", 9, 20, 10, SmartHomeUnits.VOLT),
-        CURRENT("var V  = ", 9, 20, 100, SmartHomeUnits.AMPERE),
-        POWER("var P=", 6, 20, 466, SmartHomeUnits.WATT),
-        ENERGY("var E=", 6, 20, 25600, SmartHomeUnits.WATT_HOUR);
-
-        private final Logger logger = LoggerFactory.getLogger(PWMState.class);
-
-        private final String stateResponseSearch;
-        private final int start;
-        private final int stop;
-        private final int divisor;
-        private final Unit<?> unit;
-
-        private PWMState(final String stateResponseSearch, final int start, final int stop, final int divisor,
-                final Unit<?> unit) {
-            this.stateResponseSearch = stateResponseSearch;
-            this.start = start;
-            this.stop = stop;
-            this.divisor = divisor;
-            this.unit = unit;
-        }
-
-        public String getStateResponseSearch() {
-            return stateResponseSearch;
-        }
-
-        public int getStart() {
-            return start;
-        }
-
-        public int getStop() {
-            return stop;
-        }
-
-        public int getDivisor() {
-            return divisor;
-        }
-
-        public State readState(final String loginResponseString) {
-            final int findState = loginResponseString.lastIndexOf(stateResponseSearch);
-
-            if (findState > 0) {
-                logger.trace("searchstring {} found at position {}", stateResponseSearch, findState);
-                final String slicedResponseTmp = loginResponseString.substring(findState + start, findState + stop);
-                logger.trace("transformed state response = {}", slicedResponseTmp);
-                final String[] slicedResponse = slicedResponseTmp.split(";");
-                logger.trace("transformed state response = {} - {}", slicedResponse[0], slicedResponse[1]);
-                final double value;
-
-                try {
-                    if (Double.parseDouble(slicedResponse[0]) / 1 == Double.parseDouble(slicedResponse[0])) {
-                        value = Double.parseDouble(slicedResponse[0]) / divisor;
-                    } else {
-                        value = -1.0;
-                    }
-                    return QuantityType.valueOf(value, unit);
-
-                } catch (NumberFormatException e) {
-                    logger.debug("Could not Parse State", e);
-                    return UnDefType.UNDEF;
-                }
-            } else {
-                logger.trace("searchstring '{} not found", stateResponseSearch);
-                return UnDefType.UNDEF;
-            }
-        }
-    }
 
     private final Logger logger = LoggerFactory.getLogger(EnergeniePWMHandler.class);
 
@@ -137,7 +63,7 @@ public class EnergeniePWMHandler extends BaseThingHandler {
     public void handleCommand(ChannelUID channelUID, Command command) {
         if (command instanceof RefreshType) {
             try {
-                getState();
+                scheduler.execute(this::getState);
             } catch (Exception e) {
                 logger.debug("Exception during poll", e);
             }
@@ -174,7 +100,7 @@ public class EnergeniePWMHandler extends BaseThingHandler {
         }
     }
 
-    public void getState() {
+    public synchronized void getState() {
         String url = "http://" + host + "/login.html";
         String urlParameters = "pw=" + password;
         InputStream urlContent = new ByteArrayInputStream(urlParameters.getBytes(StandardCharsets.UTF_8));
@@ -185,10 +111,10 @@ public class EnergeniePWMHandler extends BaseThingHandler {
             loginResponseString = HttpUtil.executeUrl("POST", url, urlContent, "TEXT/PLAIN", HTTP_TIMEOUT_MILLISECONDS);
 
             if (loginResponseString != null) {
-                updateState("voltage", PWMState.VOLTAGE.readState(loginResponseString));
-                updateState("current", PWMState.CURRENT.readState(loginResponseString));
-                updateState("power", PWMState.POWER.readState(loginResponseString));
-                updateState("energy", PWMState.ENERGY.readState(loginResponseString));
+                updateState("voltage", EnergeniePWMStateEnum.VOLTAGE.readState(loginResponseString));
+                updateState("current", EnergeniePWMStateEnum.CURRENT.readState(loginResponseString));
+                updateState("power", EnergeniePWMStateEnum.POWER.readState(loginResponseString));
+                updateState("energy", EnergeniePWMStateEnum.ENERGY.readState(loginResponseString));
                 try {
                     HttpUtil.executeUrl("POST", url, HTTP_TIMEOUT_MILLISECONDS);
                     logger.trace("logout from ip {}", host);
@@ -205,8 +131,9 @@ public class EnergeniePWMHandler extends BaseThingHandler {
     }
 
     private synchronized void onUpdate() {
+        ScheduledFuture<?> refreshJob = this.refreshJob;
         if (refreshJob == null || refreshJob.isCancelled()) {
-            refreshJob = scheduler.scheduleWithFixedDelay(this::getState, 5, refreshInterval, TimeUnit.SECONDS);
+            this.refreshJob = scheduler.scheduleWithFixedDelay(this::getState, 5, refreshInterval, TimeUnit.SECONDS);
         }
     }
